Fix 30 CodeQL integer conversion warnings
Add bounds checks for int64-to-int narrowing conversions across stdlib
and VM. Uses safeInt() helper for error paths, int64-space comparisons
for index bounds, and clamping for time.Date arguments.

+++ internal/vm/ops.go
@@ -89,10 +89,10 @@ func getIndex(x, idx runtime.Value) (runtime.Value, error) {
 		if err != nil {
 			return runtime.Null(), err
 		}
-		if i < 0 || int(i) >= len(lo.Items) {
+		if i < 0 || i >= int64(len(lo.Items)) {
 			return runtime.Null(), fmt.Errorf("index %d out of range (list has %d elements)", i, len(lo.Items))
 		}
-		return lo.Items[i], nil
+		return lo.Items[int(i)], nil
 	case runtime.KindMap:
 		mo := x.Obj.(*runtime.MapObj)
 		k := idx.String()
@@ -107,10 +107,10 @@ func getIndex(x, idx runtime.Value) (runtime.Value, error) {
 			return runtime.Null(), err
 		}
 		runes := []rune(x.S)
-		if i < 0 || int(i) >= len(runes) {
+		if i < 0 || i >= int64(len(runes)) {
 			return runtime.Null(), fmt.Errorf("index %d out of range (string has %d characters)", i, len(runes))
 		}
-		return runtime.Str(string(runes[i])), nil
+		return runtime.Str(string(runes[int(i)])), nil
 	default:
 		return runtime.Null(), fmt.Errorf("cannot index %s", x.KindName())
 	}
@@ -124,10 +124,10 @@ func setIndex(x, idx, val runtime.Value) error {
 		if err != nil {
 			return err
 		}
-		if i < 0 || int(i) >= len(lo.Items) {
+		if i < 0 || i >= int64(len(lo.Items)) {
 			return fmt.Errorf("index %d out of range (list has %d elements)", i, len(lo.Items))
 		}
-		lo.Items[i] = val
+		lo.Items[int(i)] = val
 		return nil
 	case runtime.KindMap:
 		mo := x.Obj.(*runtime.MapObj)
